+++ packages/scoutgame-ui/src/components/scout/components/ScoutPageCarousel/components/PromoCard.tsx
